feat: find photobooth service logics
